+++ messages/cmds/cmds.go
@@ -4,10 +4,17 @@ import (
 	"github.com/greenbone/eulabeia/messages"
 )
 
+type eventType struct{}
+
+func (eventType) Event() messages.EventType {
+	return messages.CMD
+}
+
 // Create indicates that a new entity should be created.
 // The type of of entity is indicated by `message_type`
 // e.g. "message_type": "create.target" creates a target.
 type Create struct {
+	eventType
 	messages.Message
 }
 
@@ -15,6 +22,7 @@ type Create struct {
 //
 // The response for Get is usually the aggragte with Message information and can be found within a model.
 type Get struct {
+	eventType
 	messages.Identifier
 }
 
@@ -22,6 +30,7 @@ type Get struct {
 //
 // The response of Delete is Deleted.
 type Delete struct {
+	eventType
 	messages.Identifier
 }
 
@@ -29,6 +38,7 @@ type Delete struct {
 //
 // As an example an event with the stype start.scan with the id 1 would start scan id 1
 type Start struct {
+	eventType
 	messages.Identifier
 }
 
@@ -37,6 +47,7 @@ type Start struct {
 // The values to modify are within Values, they typically match the names of the aggregate
 // but with lower case starting.
 type Modify struct {
+	eventType
 	messages.Identifier
 	Values map[string]interface{} `json:"values"`
 }